requirements.txt
dash
pandas
plotly.express
